separate composer layout components
