+++ frontend/dialob-composer-material/src/App.tsx
@@ -4,7 +4,7 @@ import { DebugFormView } from './views/DebugFormView';
 
 // TODO: Use built-in test form here
 import testForm from './dialob/test/testForm.json';
-import ComposerLayoutView from './views/ComposerLayoutView';
+import ComposerLayoutView from './views/layout/ComposerLayoutView';
 import { IntlProvider } from 'react-intl';
 import messages from './intl';
