+++ frontend/dialob-composer-material/src/views/layout/EditorArea.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const EditorArea: React.FC = () => {
+  return (
+    <>
+      Editor Area
+    </>
+  );
+};
+
+export default EditorArea;
