+++ frontend/dialob-composer-material/src/views/layout/ErrorPane.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ErrorPane: React.FC = () => {
+  return (
+    <>
+      Error Pane
+    </>
+  );
+};
+
+export default ErrorPane;
