+++ frontend/dialob-composer-material/src/views/layout/NavigationPane.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const NavigationPane: React.FC = () => {
+  return (
+    <>
+      Navigation Pane
+    </>
+  );
+};
+
+export default NavigationPane;
